+++ lib/libc/libc.csv
@@ -95,14 +95,21 @@
 "pthread_barrierattr_getpshared", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR const pthread_barrierattr_t *", "FAR int *"
 "pthread_barrierattr_init", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_barrierattr_t *"
 "pthread_barrierattr_setpshared", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_barrierattr_t *", "int"
+"pthread_barrier_destroy","pthread.h","!defined(CONFIG_DISABLE_PTHREAD)","int","FAR pthread_barrier_t*"
+"pthread_barrier_init","pthread.h","!defined(CONFIG_DISABLE_PTHREAD)","int","FAR pthread_barrier_t*","FAR const pthread_barrierattr_t*","unsigned int"
+"pthread_barrier_wait","pthread.h","!defined(CONFIG_DISABLE_PTHREAD)","int","FAR pthread_barrier_t*"
 "pthread_condattr_destroy", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_condattr_t *"
 "pthread_condattr_init", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_condattr_t *"
+"pthread_cond_destroy","pthread.h","!defined(CONFIG_DISABLE_PTHREAD)","int","FAR pthread_cond_t*"
+"pthread_cond_init","pthread.h","!defined(CONFIG_DISABLE_PTHREAD)","int","FAR pthread_cond_t*","FAR const pthread_condattr_t*"
 "pthread_mutexattr_destroy", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_mutexattr_t *"
 "pthread_mutexattr_getpshared", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_mutexattr_t *", "FAR int *"
 "pthread_mutexattr_gettype", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD) && defined(CONFIG_PTHREAD_MUTEX_TYPES)", "int", "const pthread_mutexattr_t *", "int *"
 "pthread_mutexattr_init", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_mutexattr_t *"
 "pthread_mutexattr_setpshared", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_mutexattr_t *", "int "
 "pthread_mutexattr_settype", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD) && defined(CONFIG_PTHREAD_MUTEX_TYPES)", "int", "pthread_mutexattr_t *", "int"
+"pthread_once","pthread.h","!defined(CONFIG_DISABLE_PTHREAD)","int","FAR pthread_once_t*","CODE void (*)(void)"
+"pthread_yield","pthread.h","!defined(CONFIG_DISABLE_PTHREAD)","void"
 "puts", "stdio.h", "CONFIG_NFILE_DESCRIPTORS > 0 && CONFIG_NFILE_STREAMS > 0", "int", "FAR const char *"
 "qsort", "stdlib.h", "", "void", "void *", "size_t", "size_t", "int(*)(const void *", "const void *)"
 "rand", "stdlib.h", "", "int"
