+++ lib/libc/pthread/Make.defs
@@ -61,7 +61,9 @@ CSRCS += pthread_attrsetstacksize.c pthread_attrgetstacksize.c
 CSRCS += pthread_attrsetschedparam.c pthread_attrgetschedparam.c
 CSRCS += pthread_barrierattrinit.c pthread_barrierattrdestroy.c
 CSRCS += pthread_barrierattrgetpshared.c pthread_barrierattrsetpshared.c
+CSRCS += pthread_barrierinit.c pthread_barrierdestroy.c pthread_barrierwait.c
 CSRCS += pthread_condattrinit.c pthread_condattrdestroy.c
+CSRCS += pthread_condinit.c pthread_conddestroy.c
 CSRCS += pthread_mutexattrinit.c pthread_mutexattrdestroy.c
 CSRCS += pthread_mutexattrgetpshared.c pthread_mutexattrsetpshared.c
 CSRCS += pthread_mutexattr_setprotocol.c pthread_mutexattr_getprotocol.c
@@ -69,6 +71,8 @@ CSRCS += pthread_mutexattrsettype.c pthread_mutexattrgettype.c
 CSRCS += pthread_mutexattr_setrobust.c pthread_mutexattr_getrobust.c
 CSRCS += pthread_setcancelstate.c pthread_setcanceltype.c
 CSRCS += pthread_testcancel.c
+CSRCS += pthread_rwlock.c pthread_rwlock_rdlock.c pthread_rwlock_wrlock.c
+CSRCS += pthread_once.c pthread_yield.c
 
 ifeq ($(CONFIG_ENABLE_IOTIVITY),y)
 CSRCS += pthread_condattrsetclock.c
