+++ lib/libc/pthread/pthread_barrierinit.c
@@ -16,7 +16,7 @@
  *
  ****************************************************************************/
 /********************************************************************************
- * kernel/pthread/pthread_barrieinit.c
+ * libc/pthread/pthread_barrieinit.c
  *
  *   Copyright (C) 2007, 2009 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
