+++ lib/libc/pthread/pthread_barrierwait.c
@@ -16,7 +16,7 @@
  *
  ****************************************************************************/
 /********************************************************************************
- * kernel/pthread/pthread_barrierwait.c
+ * libc/pthread/pthread_barrierwait.c
  *
  *   Copyright (C) 2007, 2009, 2014 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
