+++ lib/libc/pthread/pthread_condinit.c
@@ -16,7 +16,7 @@
  *
  ****************************************************************************/
 /****************************************************************************
- * kernel/pthread/pthread_condinit.c
+ * libc/pthread/pthread_condinit.c
  *
  *   Copyright (C) 2007-2009 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
@@ -59,8 +59,6 @@
 #include <pthread.h>
 #include <debug.h>
 #include <errno.h>
-
-#include "pthread/pthread.h"
 
 /****************************************************************************
  * Global Functions
