+++ lib/libc/pthread/pthread_yield.c
@@ -16,7 +16,7 @@
  *
  ****************************************************************************/
 /************************************************************************
- * kernel/pthread/pthread_yield.c
+ * libc/pthread/pthread_yield.c
  *
  *   Copyright (C) 2007, 2009 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
