+++ os/include/sys/syscall.h
@@ -328,42 +328,35 @@
 /* The following are defined if pthreads are enabled */
 
 #ifndef CONFIG_DISABLE_PTHREAD
-#define SYS_pthread_barrier_destroy    (__SYS_pthread+0)
-#define SYS_pthread_barrier_init       (__SYS_pthread+1)
-#define SYS_pthread_barrier_wait       (__SYS_pthread+2)
-#define SYS_pthread_cancel             (__SYS_pthread+3)
-#define SYS_pthread_cond_broadcast     (__SYS_pthread+4)
-#define SYS_pthread_cond_destroy       (__SYS_pthread+5)
-#define SYS_pthread_cond_init          (__SYS_pthread+6)
-#define SYS_pthread_cond_signal        (__SYS_pthread+7)
-#define SYS_pthread_cond_wait          (__SYS_pthread+8)
-#define SYS_pthread_create             (__SYS_pthread+9)
-#define SYS_pthread_detach             (__SYS_pthread+10)
-#define SYS_pthread_exit               (__SYS_pthread+11)
-#define SYS_pthread_getschedparam      (__SYS_pthread+12)
-#define SYS_pthread_getspecific        (__SYS_pthread+13)
-#define SYS_pthread_join               (__SYS_pthread+14)
-#define SYS_pthread_key_create         (__SYS_pthread+15)
-#define SYS_pthread_key_delete         (__SYS_pthread+16)
-#define SYS_pthread_mutex_destroy      (__SYS_pthread+17)
-#define SYS_pthread_mutex_init         (__SYS_pthread+18)
-#define SYS_pthread_mutex_lock         (__SYS_pthread+19)
-#define SYS_pthread_mutex_trylock      (__SYS_pthread+20)
-#define SYS_pthread_mutex_unlock       (__SYS_pthread+21)
-#define SYS_pthread_once               (__SYS_pthread+22)
-#define SYS_pthread_setcancelstate     (__SYS_pthread+23)
-#define SYS_pthread_setschedparam      (__SYS_pthread+24)
-#define SYS_pthread_setschedprio       (__SYS_pthread+25)
-#define SYS_pthread_setspecific        (__SYS_pthread+26)
-#define SYS_pthread_yield              (__SYS_pthread+27)
+#define SYS_pthread_cancel             (__SYS_pthread+0)
+#define SYS_pthread_cond_broadcast     (__SYS_pthread+1)
+#define SYS_pthread_cond_signal        (__SYS_pthread+2)
+#define SYS_pthread_cond_wait          (__SYS_pthread+3)
+#define SYS_pthread_create             (__SYS_pthread+4)
+#define SYS_pthread_detach             (__SYS_pthread+5)
+#define SYS_pthread_exit               (__SYS_pthread+6)
+#define SYS_pthread_getschedparam      (__SYS_pthread+7)
+#define SYS_pthread_getspecific        (__SYS_pthread+8)
+#define SYS_pthread_join               (__SYS_pthread+9)
+#define SYS_pthread_key_create         (__SYS_pthread+10)
+#define SYS_pthread_key_delete         (__SYS_pthread+11)
+#define SYS_pthread_mutex_destroy      (__SYS_pthread+12)
+#define SYS_pthread_mutex_init         (__SYS_pthread+13)
+#define SYS_pthread_mutex_lock         (__SYS_pthread+14)
+#define SYS_pthread_mutex_trylock      (__SYS_pthread+15)
+#define SYS_pthread_mutex_unlock       (__SYS_pthread+16)
+#define SYS_pthread_setcancelstate     (__SYS_pthread+17)
+#define SYS_pthread_setschedparam      (__SYS_pthread+18)
+#define SYS_pthread_setschedprio       (__SYS_pthread+19)
+#define SYS_pthread_setspecific        (__SYS_pthread+20)
 
 #ifndef CONFIG_DISABLE_SIGNAL
-#define SYS_pthread_cond_timedwait     (__SYS_pthread+28)
-#define SYS_pthread_kill               (__SYS_pthread+29)
-#define SYS_pthread_sigmask            (__SYS_pthread+30)
-#define __SYS_mqueue                   (__SYS_pthread+31)
+#define SYS_pthread_cond_timedwait     (__SYS_pthread+21)
+#define SYS_pthread_kill               (__SYS_pthread+22)
+#define SYS_pthread_sigmask            (__SYS_pthread+23)
+#define __SYS_mqueue                   (__SYS_pthread+24)
 #else
-#define __SYS_mqueue                   (__SYS_pthread+28)
+#define __SYS_mqueue                   (__SYS_pthread+21)
 #endif
 
 #else
