+++ os/syscall/syscall.csv
@@ -58,13 +58,8 @@
 "pwrite", "unistd.h", "CONFIG_NSOCKET_DESCRIPTORS > 0 || CONFIG_NFILE_DESCRIPTORS > 0", "ssize_t", "int", "FAR const void*", "size_t", "off_t"
 "posix_spawnp", "spawn.h", "defined(CONFIG_LIBC_EXECFUNCS)", "int", "FAR pid_t *", "FAR const char *", "FAR const posix_spawn_file_actions_t *", "FAR const posix_spawnattr_t *", "FAR char *const []|FAR char *const *", "FAR char *const []"
 "posix_spawn", "spawn.h", "defined(CONFIG_LIBC_EXECFUNCS)", "int", "FAR pid_t *", "FAR const char *", "FAR const posix_spawn_file_actions_t *", "FAR const posix_spawnattr_t *", "FAR char *const []|FAR char *const *", "FAR char *const []|FAR char *const *"
-"pthread_barrier_destroy", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_barrier_t*"
-"pthread_barrier_init", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_barrier_t*", "FAR const pthread_barrierattr_t*", "unsigned int"
-"pthread_barrier_wait", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_barrier_t*"
 "pthread_cancel", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "pthread_t"
 "pthread_cond_broadcast", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_cond_t*"
-"pthread_cond_destroy", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_cond_t*"
-"pthread_cond_init", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_cond_t*", "FAR const pthread_condattr_t*"
 "pthread_cond_signal", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_cond_t*"
 "pthread_cond_timedwait", "pthread.h", "!defined(CONFIG_DISABLE_SIGNALS) && !defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_cond_t*", "FAR pthread_mutex_t*", "FAR const struct timespec*"
 "pthread_cond_wait", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_cond_t*", "FAR pthread_mutex_t*"
@@ -82,13 +77,11 @@
 "pthread_mutex_lock", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_mutex_t*"
 "pthread_mutex_trylock", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_mutex_t*"
 "pthread_mutex_unlock", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_mutex_t*"
-"pthread_once", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "FAR pthread_once_t*", "CODE void (*)(void)"
 "pthread_setcancelstate", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "int", "FAR int*"
 "pthread_setschedparam", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "pthread_t", "int", "FAR const struct sched_param*"
 "pthread_setschedprio", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "pthread_t", "int"
 "pthread_setspecific", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "int", "pthread_key_t", "FAR const void*"
 "pthread_sigmask", "pthread.h", "!defined(CONFIG_DISABLE_SIGNALS) && !defined(CONFIG_DISABLE_PTHREAD)", "int", "int", "FAR const sigset_t*", "FAR sigset_t*"
-"pthread_yield", "pthread.h", "!defined(CONFIG_DISABLE_PTHREAD)", "void"
 "putenv", "stdlib.h", "!defined(CONFIG_DISABLE_ENVIRON)", "int", "FAR const char*"
 "read", "unistd.h", "CONFIG_NSOCKET_DESCRIPTORS > 0 || CONFIG_NFILE_DESCRIPTORS > 0", "ssize_t", "int", "FAR void*", "size_t"
 "readdir", "dirent.h", "CONFIG_NFILE_DESCRIPTORS > 0", "FAR struct dirent*", "FAR DIR*"
