+++ os/syscall/syscall_lookup.h
@@ -247,13 +247,8 @@ SYSCALL_LOOKUP(shmdt,                   1, STUB_shmdt)
 /* The following are defined if pthreads are enabled */
 
 #ifndef CONFIG_DISABLE_PTHREAD
-SYSCALL_LOOKUP(pthread_barrier_destroy, 1, STUB_pthread_barrier_destroy)
-SYSCALL_LOOKUP(pthread_barrier_init,    3, STUB_pthread_barrier_init)
-SYSCALL_LOOKUP(pthread_barrier_wait,    1, STUB_pthread_barrier_wait)
 SYSCALL_LOOKUP(pthread_cancel,          1, STUB_pthread_cancel)
 SYSCALL_LOOKUP(pthread_cond_broadcast,  1, STUB_pthread_cond_broadcast)
-SYSCALL_LOOKUP(pthread_cond_destroy,    1, STUB_pthread_cond_destroy)
-SYSCALL_LOOKUP(pthread_cond_init,       2, STUB_pthread_cond_init)
 SYSCALL_LOOKUP(pthread_cond_signal,     1, STUB_pthread_cond_signal)
 SYSCALL_LOOKUP(pthread_cond_wait,       2, STUB_pthread_cond_wait)
 SYSCALL_LOOKUP(pthread_create,          4, STUB_pthread_create)
@@ -269,12 +264,10 @@ SYSCALL_LOOKUP(pthread_mutex_init,      2, STUB_pthread_mutex_init)
 SYSCALL_LOOKUP(pthread_mutex_lock,      1, STUB_pthread_mutex_lock)
 SYSCALL_LOOKUP(pthread_mutex_trylock,   1, STUB_pthread_mutex_trylock)
 SYSCALL_LOOKUP(pthread_mutex_unlock,    1, STUB_pthread_mutex_unlock)
-SYSCALL_LOOKUP(pthread_once,            2, STUB_pthread_once)
 SYSCALL_LOOKUP(pthread_setcancelstate,  2, STUB_pthread_setcancelstate)
 SYSCALL_LOOKUP(pthread_setschedparam,   3, STUB_pthread_setschedparam)
 SYSCALL_LOOKUP(pthread_setschedprio,    2, STUB_pthread_setschedprio)
 SYSCALL_LOOKUP(pthread_setspecific,     2, STUB_pthread_setspecific)
-SYSCALL_LOOKUP(pthread_yield,           0, STUB_pthread_yield)
 #  ifndef CONFIG_DISABLE_SIGNAL
 SYSCALL_LOOKUP(pthread_cond_timedwait,  3, STUB_pthread_cond_timedwait)
 SYSCALL_LOOKUP(pthread_kill,            2, STUB_pthread_kill)
